Fix infinite loop when redo-always && redo-stamp a non existant target
Added test case where a target b is redo-always and redo-stamp</dev/null and
b is never written on disk. This caused redo to think b bas never build even
though it had a stamp and its regular stamp included a runid. The fix makes
prevents redo from thinking the target was never build if it has a runid in
its stamp.

The test is not supposed to fail but might trigger in infinite loop.