BUG: Fix redo-ood, redo-sources, and redo-targets
When no targets are specified the function init(..) was making 'all'
the default target.  However, this caused these redo flavours to fail because
they require that no targets be specified when they are invoked.  This
fix makes 'all' as the default target for only 'redo' and 'redo-ifchange'.
It also fixes an outdated function that searched for '*.deps.redo' files
instead of '.redo/*.deps'.