Added redo-dofile executable
redo-dofile is a redo executable that prints the .do files that would build
the targets passed as arguments. If any target can't be built (no
corresponding .do file was found) then the command return a failure status.
This can be very useful to introspect into redo and find if a target is
buildable.